Flush recorder after tests runs so it can be initialised globally
This is to support writing from a database driver, which must only be
initialised once
