+++ examples/sequence-diagrams-with-postgres-database/Makefile
@@ -1,3 +1,5 @@
+.PHONY: postgres-dev test
+
 postgres-dev:
 	docker stop apitest_postgres || true && docker rm apitest_postgres || true
 	docker run -d -e POSTGRES_PASSWORD=postgres -e POSTGRES_USER=postgres -e POSTGRES_DB=apitest -p 5432:5432 --name apitest_postgres postgres:9.6
