+++ examples/sequence-diagrams-with-sqlite-database/Makefile
@@ -1,2 +1,4 @@
+.PHONY: test
+
 test:
 	SQLITE_DSN="./foo.db" go test ./... -v
